+++ packages/node/CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.6.6] - 2024-06-21
+### Fixed
+- Update with common-substrate, fix `EventFilter` incorrectly extend `BlockFilter`, lead dictionary error (#2463)
+- Fix dictionary query entries included undefined fields (#2463)
+
 ## [4.6.5] - 2024-06-18
 ### Fixed
 - Add missing chainType validation for local mode
@@ -1228,7 +1233,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - bump @polkadot/api to 3.1.1
 
-[Unreleased]: https://github.com/subquery/subql/compare/node/4.6.5...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/node/4.6.6...HEAD
+[4.6.6]: https://github.com/subquery/subql/compare/node/4.6.5...node/4.6.6
 [4.6.5]: https://github.com/subquery/subql/compare/node/4.6.4...node/4.6.5
 [4.6.4]: https://github.com/subquery/subql/compare/node/4.6.3...node/4.6.4
 [4.6.3]: https://github.com/subquery/subql/compare/node/4.6.2...node/4.6.3